requirements.txt
absl-py==0.6.1
astor==0.7.1
backcall==0.1.0
bleach==3.0.2
certifi==2018.10.15
chardet==3.0.4
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
gast==0.2.0
grpcio==1.16.1
h5py==2.8.0
idna==2.7
imageio==2.4.1
ipdb==0.11
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonlines==1.2.0
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
MarkupSafe==1.1.0
mistune==0.8.4
mock==2.0.0
moviepy==0.2.3.5
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.2
numpy==1.15.4
opencv-python==3.4.3.18
pandas==0.23.4
pandocfilters==1.4.2
parse==1.9.0
parso==0.3.1
pbr==5.1.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.2.0
python-dateutil==2.7.5
pytz==2018.7
pyzmq==17.1.2
requests==2.20.1
scipy==1.1.0
Send2Trash==1.5.0
six==1.11.0
tb-nightly==1.13.0a20181118
tensorboard==1.12.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tf-estimator-nightly==1.12.0.dev20181119
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
